requirements.txt
Flask==3.0.2
spotipy==2.23.0
gunicorn==21.2.0
python-dotenv==1.0.1
Werkzeug==3.0.1 